submissions/nikeshboopath-jpg/requirements.txt
requests>=2.0
pytest>=6.0
python-dotenv>=1.0
